third_party/WebKit/LayoutTests/fast/css/variables/cssom-read.html

<!doctype html>
<head><script src="../../js/resources/js-test-pre.js"></script></head>

<div id="test-target" style="var-a: override me; var-a: apple; var-b: banana;"></div>

<script>
description('This tests Javascript read access to CSS variables in inline styles.');

var div = document.querySelector('#test-target');

shouldBeEqualToString('div.style.var.toString()', '[object CSSVariablesMap]');
shouldBe('div.style.var === div.style.var', 'true');
shouldBeEqualToString('div.style.cssText', 'var-a: apple; var-b: banana;');
shouldBe('div.style.var.size', '2');
shouldBeEqualToString('div.style.var.get("a")', 'apple');
shouldBeEqualToString('div.style.var.get("b")', 'banana');
shouldBeEqualToString('div.style.var.get("nonexistant")', '');

</script>
<script src="../../js/resources/js-test-post.js"></script>
